feat: configured pages to setup the layout by group routing
