+++ apps/bytebank/src/app/(pages)/investimentos/page.tsx
@@ -1,0 +1,6 @@
+export default function Investimentos() {
+
+  return (
+      <p>Investimentos</p>
+  );
+}
