+++ apps/bytebank/src/app/(pages)/outros/page.tsx
@@ -1,0 +1,4 @@
+export default function Outros() {
+
+  return (<p>Outros</p>);
+}
