+++ apps/bytebank/src/app/(pages)/transferencias/page.tsx
@@ -1,0 +1,4 @@
+export default function Transferencias() {
+
+  return (<p>Transferencias</p>);
+}
