--- apps/bytebank/src/app/investimentos/page.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import { BytebankProvider } from "@bytebank/shared";
-
-export default function Investimentos() {
-
-  return (
-    <BytebankProvider>
-      Investimentos
-    </BytebankProvider>
-  );
-}
