+++ apps/bytebank/src/app/not-found.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { BytebankProvider, BytebankIllustration, BytebankButton, BytebankText, palette } from "@bytebank/shared";
+import { BytebankIllustration, BytebankButton, BytebankText, palette, BytebankProvider as Provider } from "@bytebank/shared";
 import { Box } from "@mui/material";
 import { useRouter } from "next/navigation";
 
@@ -10,18 +10,18 @@ export default function NotFound() {
 
   return (
     <>
-      <BytebankProvider canNavigate={false}>
+      <Provider.BytebankProvider canNavigate={false}>
         <Box display={"flex"} flexDirection={"column"} alignItems={"center"} justifyContent={"center"} flexGrow={1} gap={2}>
           <BytebankIllustration name={"error"} />
 
           <BytebankText variant={"lg"} color={palette["grey.900"]}> Ops! Não encontramos a página... </BytebankText>
-          <BytebankText variant={"sm"} color={palette["grey.900"]}>  E olha que exploramos o universo procurando por ela! 
-Que tal voltar e tentar novamente?</BytebankText>
+          <BytebankText variant={"sm"} color={palette["grey.900"]}>  E olha que exploramos o universo procurando por ela!
+            Que tal voltar e tentar novamente?</BytebankText>
           <Box marginTop={4}>
-          <BytebankButton label="Voltar ao início" variant={"contained"} color={"secondary"} sendSubmit={() => handleRedirect()} />
+            <BytebankButton label="Voltar ao início" variant={"contained"} color={"secondary"} sendSubmit={() => handleRedirect()} />
           </Box>
         </Box>
-      </BytebankProvider>
+      </Provider.BytebankProvider>
     </>
   );
 }
