--- apps/bytebank/src/app/outros/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-import { BytebankProvider } from "@bytebank/shared";
-
-export default function Outros() {
-
-  return (
-    <BytebankProvider>Outros</BytebankProvider>
-  );
-}
