+++ apps/bytebank/src/app/page.tsx
@@ -5,7 +5,6 @@ import {
   BytebankModal,
   BytebankSelectController,
   BytebankInputController,
-  BytebankProvider,
 } from '@bytebank/shared';
 import { useForm, FormProvider } from 'react-hook-form';
 import { Box } from '@mui/material';
@@ -44,58 +43,55 @@ export default function Index() {
 
   return (
     <>
-      <BytebankProvider>
-        <Box width="100%" p={2} display="flex" flexDirection="column" gap={4}>
-          <BytebankCardBank
-            variant="physical"
-            details={cardDetails}
-          ></BytebankCardBank>
-          <BytebankCardBank
-            variant="virtual"
-            details={cardDetails}
-          ></BytebankCardBank>
-          <BytebankButton
-            sendSubmit={() => setOpen(true)}
-            label="Abrir Modal"
-            color="primary"
-            variant="outlined" />
-        </Box><BytebankModal
-          illustrationSize="lg"
-          title="Preencha os campos abaixo para criar sua conta corrente!"
-          illustration="register"
-          open={open}
-          onClose={() => setOpen(false)}
-        >
-          <>
-            <FormProvider {...methods}>
-              <form onSubmit={methods.handleSubmit(onSubmit)} noValidate>
-                <BytebankInputController
-                  name="name"
-                  label="Nome"
-                  placeholder="Digite seu nome" />
-                <BytebankInputController
-                  name="email"
-                  label="Email"
-                  placeholder="Digite seu email"
-                  type="email" />
+      <Box width="100%" p={2} display="flex" flexDirection="column" gap={4}>
+        <BytebankCardBank
+          variant="physical"
+          details={cardDetails}
+        ></BytebankCardBank>
+        <BytebankCardBank
+          variant="virtual"
+          details={cardDetails}
+        ></BytebankCardBank>
+        <BytebankButton
+          sendSubmit={() => setOpen(true)}
+          label="Abrir Modal"
+          color="primary"
+          variant="outlined" />
+      </Box><BytebankModal
+        illustrationSize="lg"
+        title="Preencha os campos abaixo para criar sua conta corrente!"
+        illustration="register"
+        open={open}
+        onClose={() => setOpen(false)}
+      >
+        <>
+          <FormProvider {...methods}>
+            <form onSubmit={methods.handleSubmit(onSubmit)} noValidate>
+              <BytebankInputController
+                name="name"
+                label="Nome"
+                placeholder="Digite seu nome" />
+              <BytebankInputController
+                name="email"
+                label="Email"
+                placeholder="Digite seu email"
+                type="email" />
 
-                <BytebankSelectController
-                  name="tipo"
-                  label="Tipo de pessoa"
-                  options={selectOptions} />
+              <BytebankSelectController
+                name="tipo"
+                label="Tipo de pessoa"
+                options={selectOptions} />
 
-                <Box>
-                  <BytebankButton
-                    label="Concluir transação"
-                    color="primary"
-                    variant="contained" />
-                </Box>
-              </form>
-            </FormProvider>
-          </>
-        </BytebankModal>
-      </BytebankProvider>
-
+              <Box>
+                <BytebankButton
+                  label="Concluir transação"
+                  color="primary"
+                  variant="contained" />
+              </Box>
+            </form>
+          </FormProvider>
+        </>
+      </BytebankModal>
     </>
   );
 }
