--- apps/bytebank/src/app/transferencias/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-import { BytebankProvider } from "@bytebank/shared";
-
-export default function Transferencias() {
-
-  return (
-    <BytebankProvider>Transferencias</BytebankProvider>
-  );
-}
