+++ package.json
@@ -15,7 +15,7 @@
     "@mui/icons-material": "^7.0.2",
     "@mui/material": "^7.0.2",
     "@toolpad/core": "^0.13.0",
-    "next": "~15.1.4",
+    "next": "^15.3.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-hook-form": "^7.55.0"
